Hostname change and transition engine fix
Tried to set the hostname to something useful, both with the WiFi
library and ESP8266 library, but it doesn't seem to actually work.  Oh
well.

Fixed a problem where the transition time was being stored before
sending the HTTP request to SmartThings.  The request takes just long
enough to affect the actual transition, and while not noticable on power
on (due to the extra delay from the AC power coming on first), it would
cause the first two lights to seem to turn off in rapid successsion.
Simply waiting until after sending the HTTP request before doing  any
part of the transition fixes the problem.